[drawer] Co-locate and expand virtual keyboard provider tests
