+++ packages/react/src/drawer/viewport/DrawerViewport.test.tsx
@@ -47,103 +47,6 @@ describe('<Drawer.Viewport />', () => {
     return touchMove;
   }
 
-  function createNativeTouchEnd(target: EventTarget, point: { clientX: number; clientY: number }) {
-    const touchEnd = new Event('touchend', { bubbles: true, cancelable: true });
-    Object.defineProperty(touchEnd, 'changedTouches', {
-      value: [createTouch(target, point)],
-      configurable: true,
-    });
-    Object.defineProperty(touchEnd, 'touches', {
-      value: [],
-      configurable: true,
-    });
-    return touchEnd;
-  }
-
-  function createNativeTouchCancel(
-    target: EventTarget,
-    point: { clientX: number; clientY: number },
-  ) {
-    const touchCancel = new Event('touchcancel', { bubbles: true, cancelable: true });
-    Object.defineProperty(touchCancel, 'changedTouches', {
-      value: [createTouch(target, point)],
-      configurable: true,
-    });
-    Object.defineProperty(touchCancel, 'touches', {
-      value: [],
-      configurable: true,
-    });
-    return touchCancel;
-  }
-
-  function mockVisualViewport(height: number) {
-    const originalDescriptor = Object.getOwnPropertyDescriptor(window, 'visualViewport');
-    const listeners = new Map<string, Set<EventListener>>();
-
-    const visualViewport: Pick<
-      VisualViewport,
-      'addEventListener' | 'removeEventListener' | 'offsetTop'
-    > & {
-      height: number;
-      scale: number;
-    } = {
-      height,
-      offsetTop: 0,
-      scale: 1,
-      addEventListener(type: string, listener: EventListener) {
-        if (!listeners.has(type)) {
-          listeners.set(type, new Set());
-        }
-
-        listeners.get(type)?.add(listener);
-      },
-      removeEventListener(type: string, listener: EventListener) {
-        listeners.get(type)?.delete(listener);
-      },
-    };
-
-    Object.defineProperty(window, 'visualViewport', {
-      configurable: true,
-      value: visualViewport as VisualViewport,
-    });
-
-    return {
-      restore() {
-        if (originalDescriptor) {
-          Object.defineProperty(window, 'visualViewport', originalDescriptor);
-        } else {
-          Object.defineProperty(window, 'visualViewport', {
-            configurable: true,
-            value: undefined,
-          });
-        }
-      },
-      resize(nextHeight: number) {
-        visualViewport.height = nextHeight;
-        listeners.get('resize')?.forEach((listener) => listener(new Event('resize')));
-      },
-      setScale(nextScale: number) {
-        visualViewport.scale = nextScale;
-        listeners.get('resize')?.forEach((listener) => listener(new Event('resize')));
-      },
-    };
-  }
-
-  function mockWindowInnerHeight(innerHeight: number) {
-    const originalDescriptor = Object.getOwnPropertyDescriptor(window, 'innerHeight');
-
-    Object.defineProperty(window, 'innerHeight', {
-      configurable: true,
-      value: innerHeight,
-    });
-
-    return () => {
-      if (originalDescriptor) {
-        Object.defineProperty(window, 'innerHeight', originalDescriptor);
-      }
-    };
-  }
-
   it('clears text selection on swipe start', async () => {
     await render(
       <Drawer.Root open>
@@ -355,809 +258,6 @@ describe('<Drawer.Viewport />', () => {
       expect(backdrop).toHaveAttribute('data-swiping', '');
     } finally {
       document.elementFromPoint = originalElementFromPoint;
-    }
-  });
-
-  it.skipIf(isJSDOM)(
-    'adds scroll slack and centers the focused input while the visual viewport is reduced',
-    async () => {
-      const restoreInnerHeight = mockWindowInnerHeight(800);
-      const visualViewport = mockVisualViewport(800);
-
-      try {
-        await render(
-          <Drawer.Root open modal={false}>
-            <Drawer.VirtualKeyboardProvider>
-              <Drawer.Portal>
-                <Drawer.Viewport>
-                  <Drawer.Popup>
-                    <Drawer.Content
-                      data-testid="scroll"
-                      style={{
-                        height: 420,
-                        overflowY: 'auto',
-                        overflowAnchor: 'auto',
-                        paddingBottom: 20,
-                      }}
-                    >
-                      <div style={{ height: 900 }} />
-                      <input data-testid="input" type="text" />
-                    </Drawer.Content>
-                  </Drawer.Popup>
-                </Drawer.Viewport>
-              </Drawer.Portal>
-            </Drawer.VirtualKeyboardProvider>
-          </Drawer.Root>,
-        );
-
-        const scroll = screen.getByTestId('scroll');
-        const input = screen.getByTestId('input');
-
-        Object.defineProperties(scroll, {
-          clientHeight: { configurable: true, value: 420 },
-          scrollHeight: { configurable: true, value: 1200 },
-        });
-        scroll.getBoundingClientRect = () =>
-          ({
-            top: 300,
-            bottom: 720,
-            height: 420,
-            left: 0,
-            right: 320,
-            width: 320,
-            x: 0,
-            y: 300,
-            toJSON: () => {},
-          }) as DOMRect;
-        input.getBoundingClientRect = () => {
-          const top = 650 - scroll.scrollTop;
-          return {
-            top,
-            bottom: top + 40,
-            height: 40,
-            left: 0,
-            right: 320,
-            width: 320,
-            x: 0,
-            y: top,
-            toJSON: () => {},
-          } as DOMRect;
-        };
-
-        const keyboardViewportHeight = 500;
-        await act(async () => {
-          input.focus();
-          scroll.scrollTop = 0;
-          visualViewport.resize(keyboardViewportHeight);
-        });
-
-        await waitFor(() => {
-          const scrollRect = scroll.getBoundingClientRect();
-          const inputRect = input.getBoundingClientRect();
-          const inputCenter = (inputRect.top + inputRect.bottom) / 2;
-          const visibleCenter = (scrollRect.top + keyboardViewportHeight) / 2;
-
-          expect(Number.parseFloat(scroll.style.paddingBottom)).toBeGreaterThan(20);
-          expect(scroll.style.scrollPaddingBottom).not.toBe('');
-          expect(scroll.style.overflowAnchor).toBe('none');
-          expect(inputCenter).toBeCloseTo(visibleCenter, 0);
-        });
-
-        await act(async () => {
-          input.blur();
-        });
-
-        await waitFor(() => {
-          expect(scroll.style.paddingBottom).toBe('20px');
-          expect(scroll.style.scrollPaddingBottom).toBe('');
-          expect(scroll.style.overflowAnchor).toBe('auto');
-        });
-      } finally {
-        visualViewport.restore();
-        restoreInnerHeight();
-      }
-    },
-  );
-
-  it.skipIf(isJSDOM)(
-    'adds keyboard scroll slack to a potential scroll ancestor for textareas',
-    async () => {
-      const restoreInnerHeight = mockWindowInnerHeight(800);
-      const visualViewport = mockVisualViewport(800);
-
-      try {
-        await render(
-          <Drawer.Root open modal={false}>
-            <Drawer.VirtualKeyboardProvider>
-              <Drawer.Portal>
-                <Drawer.Viewport>
-                  <Drawer.Popup>
-                    <div
-                      data-testid="scroll"
-                      style={{
-                        height: 420,
-                        overflowY: 'auto',
-                        paddingBottom: 20,
-                      }}
-                    >
-                      <div style={{ height: 300 }} />
-                      <textarea data-testid="textarea" />
-                    </div>
-                  </Drawer.Popup>
-                </Drawer.Viewport>
-              </Drawer.Portal>
-            </Drawer.VirtualKeyboardProvider>
-          </Drawer.Root>,
-        );
-
-        const scroll = screen.getByTestId('scroll');
-        const textarea = screen.getByTestId('textarea');
-
-        Object.defineProperties(scroll, {
-          clientHeight: { configurable: true, value: 420 },
-          scrollHeight: { configurable: true, value: 420 },
-        });
-        Object.defineProperties(textarea, {
-          clientHeight: { configurable: true, value: 88 },
-          scrollHeight: { configurable: true, value: 88 },
-        });
-        scroll.getBoundingClientRect = () =>
-          ({
-            top: 300,
-            bottom: 720,
-            height: 420,
-            left: 0,
-            right: 320,
-            width: 320,
-            x: 0,
-            y: 300,
-            toJSON: () => {},
-          }) as DOMRect;
-        textarea.getBoundingClientRect = () =>
-          ({
-            top: 650,
-            bottom: 690,
-            height: 40,
-            left: 0,
-            right: 320,
-            width: 320,
-            x: 0,
-            y: 650,
-            toJSON: () => {},
-          }) as DOMRect;
-
-        await act(async () => {
-          textarea.focus();
-          visualViewport.resize(500);
-        });
-
-        await waitFor(() => {
-          expect(Number.parseFloat(scroll.style.paddingBottom)).toBeGreaterThan(20);
-          expect(textarea.style.paddingBottom).toBe('');
-        });
-      } finally {
-        visualViewport.restore();
-        restoreInnerHeight();
-      }
-    },
-  );
-
-  it.skipIf(isJSDOM)(
-    'sets the keyboard inset CSS variable for a focused input without a scroll target',
-    async () => {
-      const restoreInnerHeight = mockWindowInnerHeight(800);
-      const visualViewport = mockVisualViewport(800);
-
-      try {
-        await render(
-          <Drawer.Root open modal={false}>
-            <Drawer.VirtualKeyboardProvider>
-              <Drawer.Portal>
-                <Drawer.Viewport data-testid="viewport">
-                  <Drawer.Popup data-testid="popup">
-                    <div>
-                      <input data-testid="input" type="text" />
-                    </div>
-                  </Drawer.Popup>
-                </Drawer.Viewport>
-              </Drawer.Portal>
-            </Drawer.VirtualKeyboardProvider>
-          </Drawer.Root>,
-        );
-
-        const viewport = screen.getByTestId('viewport');
-        const popup = screen.getByTestId('popup');
-        const input = screen.getByTestId('input');
-
-        await act(async () => {
-          input.focus();
-          visualViewport.resize(500);
-        });
-
-        await waitFor(() => {
-          expect(viewport.style.getPropertyValue('--drawer-keyboard-inset')).toBe('300px');
-          expect(getComputedStyle(popup).getPropertyValue('--drawer-keyboard-inset')).toBe('300px');
-        });
-
-        await act(async () => {
-          input.blur();
-        });
-
-        await waitFor(() => {
-          expect(viewport.style.getPropertyValue('--drawer-keyboard-inset')).toBe('0px');
-        });
-      } finally {
-        visualViewport.restore();
-        restoreInnerHeight();
-      }
-    },
-  );
-
-  it.skipIf(isJSDOM)('does not add keyboard scroll slack by default', async () => {
-    const restoreInnerHeight = mockWindowInnerHeight(800);
-    const visualViewport = mockVisualViewport(800);
-
-    try {
-      await render(
-        <Drawer.Root open modal={false}>
-          <Drawer.Portal>
-            <Drawer.Viewport>
-              <Drawer.Popup>
-                <Drawer.Content
-                  data-testid="scroll"
-                  style={{ height: 420, overflowY: 'auto', paddingBottom: 20 }}
-                >
-                  <div style={{ height: 900 }} />
-                  <input data-testid="input" type="text" />
-                </Drawer.Content>
-              </Drawer.Popup>
-            </Drawer.Viewport>
-          </Drawer.Portal>
-        </Drawer.Root>,
-      );
-
-      const scroll = screen.getByTestId('scroll');
-      const input = screen.getByTestId('input');
-
-      await act(async () => {
-        input.focus();
-        scroll.scrollTop = 0;
-        visualViewport.resize(500);
-      });
-
-      await waitFor(() => {
-        expect(scroll.style.paddingBottom).toBe('20px');
-        expect(scroll.style.scrollPaddingBottom).toBe('');
-        expect(scroll.scrollTop).toBe(0);
-      });
-    } finally {
-      visualViewport.restore();
-      restoreInnerHeight();
-    }
-  });
-
-  it.skipIf(isJSDOM)('restores keyboard scroll slack when the drawer closes', async () => {
-    const restoreInnerHeight = mockWindowInnerHeight(800);
-    const visualViewport = mockVisualViewport(800);
-
-    function TestCase(props: { open: boolean }) {
-      return (
-        <Drawer.Root open={props.open} modal={false}>
-          <Drawer.VirtualKeyboardProvider>
-            <Drawer.Portal>
-              <Drawer.Viewport>
-                <Drawer.Popup>
-                  <Drawer.Content
-                    data-testid="scroll"
-                    style={{
-                      height: 420,
-                      overflowY: 'auto',
-                      overflowAnchor: 'auto',
-                      paddingBottom: 20,
-                    }}
-                  >
-                    <div style={{ height: 900 }} />
-                    <input data-testid="input" type="text" />
-                  </Drawer.Content>
-                </Drawer.Popup>
-              </Drawer.Viewport>
-            </Drawer.Portal>
-          </Drawer.VirtualKeyboardProvider>
-        </Drawer.Root>
-      );
-    }
-
-    try {
-      const { setProps } = await render(<TestCase open />);
-
-      const scroll = screen.getByTestId('scroll');
-      const input = screen.getByTestId('input');
-
-      Object.defineProperties(scroll, {
-        clientHeight: { configurable: true, value: 420 },
-        scrollHeight: { configurable: true, value: 1200 },
-      });
-      scroll.getBoundingClientRect = () =>
-        ({
-          top: 300,
-          bottom: 720,
-          height: 420,
-          left: 0,
-          right: 320,
-          width: 320,
-          x: 0,
-          y: 300,
-          toJSON: () => {},
-        }) as DOMRect;
-      input.getBoundingClientRect = () =>
-        ({
-          top: 650,
-          bottom: 690,
-          height: 40,
-          left: 0,
-          right: 320,
-          width: 320,
-          x: 0,
-          y: 650,
-          toJSON: () => {},
-        }) as DOMRect;
-
-      await act(async () => {
-        input.focus();
-        visualViewport.resize(500);
-      });
-
-      await waitFor(() => {
-        expect(Number.parseFloat(scroll.style.paddingBottom)).toBeGreaterThan(20);
-        expect(scroll.style.overflowAnchor).toBe('none');
-      });
-
-      await setProps({ open: false });
-
-      await waitFor(() => {
-        expect(scroll.style.paddingBottom).toBe('20px');
-        expect(scroll.style.scrollPaddingBottom).toBe('');
-        expect(scroll.style.overflowAnchor).toBe('auto');
-      });
-    } finally {
-      visualViewport.restore();
-      restoreInnerHeight();
-    }
-  });
-
-  it.skipIf(isJSDOM)('does not add keyboard scroll slack while pinch-zoomed', async () => {
-    const restoreInnerHeight = mockWindowInnerHeight(800);
-    const visualViewport = mockVisualViewport(800);
-
-    try {
-      await render(
-        <Drawer.Root open modal={false}>
-          <Drawer.VirtualKeyboardProvider>
-            <Drawer.Portal>
-              <Drawer.Viewport>
-                <Drawer.Popup>
-                  <Drawer.Content
-                    data-testid="scroll"
-                    style={{ height: 420, overflowY: 'auto', paddingBottom: 20 }}
-                  >
-                    <div style={{ height: 900 }} />
-                    <input data-testid="input" type="text" />
-                  </Drawer.Content>
-                </Drawer.Popup>
-              </Drawer.Viewport>
-            </Drawer.Portal>
-          </Drawer.VirtualKeyboardProvider>
-        </Drawer.Root>,
-      );
-
-      const scroll = screen.getByTestId('scroll');
-      const input = screen.getByTestId('input');
-
-      await act(async () => {
-        input.focus();
-        scroll.scrollTop = 0;
-        visualViewport.setScale(1.5);
-        visualViewport.resize(500);
-      });
-
-      await waitFor(() => {
-        expect(scroll.style.paddingBottom).toBe('20px');
-        expect(scroll.style.scrollPaddingBottom).toBe('');
-        expect(scroll.scrollTop).toBe(0);
-      });
-    } finally {
-      visualViewport.restore();
-      restoreInnerHeight();
-    }
-  });
-
-  it.skipIf(isJSDOM)(
-    'focuses an unfocused keyboard input on touchend without page scroll',
-    async () => {
-      await render(
-        <Drawer.Root open modal={false}>
-          <Drawer.VirtualKeyboardProvider>
-            <Drawer.Portal>
-              <Drawer.Backdrop data-testid="backdrop" />
-              <Drawer.Viewport>
-                <Drawer.Popup>
-                  <input
-                    data-testid="input"
-                    style={{
-                      opacity: 0.5,
-                      transform: 'scale(1)',
-                      transition: 'opacity 1s',
-                    }}
-                    type="text"
-                  />
-                </Drawer.Popup>
-              </Drawer.Viewport>
-            </Drawer.Portal>
-          </Drawer.VirtualKeyboardProvider>
-        </Drawer.Root>,
-      );
-
-      const backdrop = screen.getByTestId('backdrop');
-      const input = screen.getByTestId('input');
-      const focusSpy = vi.spyOn(input, 'focus');
-      const originalElementFromPoint = document.elementFromPoint;
-      document.elementFromPoint = () => input;
-
-      try {
-        fireEvent.touchStart(input, {
-          touches: [
-            createTouch(input, {
-              clientX: 0,
-              clientY: 0,
-            }),
-          ],
-        });
-
-        await waitFor(() => {
-          expect(backdrop).toHaveAttribute('data-swiping', '');
-        });
-
-        const touchEnd = createNativeTouchEnd(input, {
-          clientX: 0,
-          clientY: 0,
-        });
-
-        await act(async () => {
-          input.dispatchEvent(touchEnd);
-          await flushMicrotasks();
-        });
-
-        expect(touchEnd.defaultPrevented).toBe(true);
-        expect(focusSpy).toHaveBeenCalledWith({ preventScroll: true });
-        expect(input.style.opacity).toBe('0.5');
-        expect(input.style.transform).toBe('scale(1)');
-        expect(input.style.transition).toBe('opacity 1s');
-
-        await waitFor(() => {
-          expect(backdrop).not.toHaveAttribute('data-swiping');
-        });
-      } finally {
-        document.elementFromPoint = originalElementFromPoint;
-        focusSpy.mockRestore();
-      }
-    },
-  );
-
-  it.skipIf(isJSDOM)('preserves native taps on picker inputs', async () => {
-    await render(
-      <Drawer.Root open modal={false}>
-        <Drawer.VirtualKeyboardProvider>
-          <Drawer.Portal>
-            <Drawer.Viewport>
-              <Drawer.Popup>
-                <input data-testid="input" type="date" />
-              </Drawer.Popup>
-            </Drawer.Viewport>
-          </Drawer.Portal>
-        </Drawer.VirtualKeyboardProvider>
-      </Drawer.Root>,
-    );
-
-    const input = screen.getByTestId('input');
-    const focusSpy = vi.spyOn(input, 'focus');
-    const originalElementFromPoint = document.elementFromPoint;
-    document.elementFromPoint = () => input;
-
-    try {
-      fireEvent.touchStart(input, {
-        touches: [
-          createTouch(input, {
-            clientX: 0,
-            clientY: 0,
-          }),
-        ],
-      });
-
-      const touchEnd = createNativeTouchEnd(input, {
-        clientX: 0,
-        clientY: 0,
-      });
-
-      await act(async () => {
-        input.dispatchEvent(touchEnd);
-        await flushMicrotasks();
-      });
-
-      expect(touchEnd.defaultPrevented).toBe(false);
-      expect(focusSpy).not.toHaveBeenCalled();
-    } finally {
-      document.elementFromPoint = originalElementFromPoint;
-      focusSpy.mockRestore();
-    }
-  });
-
-  it.skipIf(isJSDOM)(
-    'focuses a keyboard input with data-base-ui-swipe-ignore without starting drawer swipe',
-    async () => {
-      await render(
-        <Drawer.Root open modal={false}>
-          <Drawer.VirtualKeyboardProvider>
-            <Drawer.Portal>
-              <Drawer.Backdrop data-testid="backdrop" />
-              <Drawer.Viewport>
-                <Drawer.Popup>
-                  <input data-testid="input" data-base-ui-swipe-ignore type="text" />
-                </Drawer.Popup>
-              </Drawer.Viewport>
-            </Drawer.Portal>
-          </Drawer.VirtualKeyboardProvider>
-        </Drawer.Root>,
-      );
-
-      const backdrop = screen.getByTestId('backdrop');
-      const input = screen.getByTestId('input');
-      const focusSpy = vi.spyOn(input, 'focus');
-      const originalElementFromPoint = document.elementFromPoint;
-      document.elementFromPoint = () => input;
-
-      try {
-        fireEvent.touchStart(input, {
-          touches: [
-            createTouch(input, {
-              clientX: 0,
-              clientY: 0,
-            }),
-          ],
-        });
-
-        await flushMicrotasks();
-
-        expect(backdrop).not.toHaveAttribute('data-swiping');
-
-        const touchEnd = createNativeTouchEnd(input, {
-          clientX: 0,
-          clientY: 0,
-        });
-
-        await act(async () => {
-          input.dispatchEvent(touchEnd);
-          await flushMicrotasks();
-        });
-
-        expect(touchEnd.defaultPrevented).toBe(true);
-        expect(focusSpy).toHaveBeenCalledWith({ preventScroll: true });
-        expect(backdrop).not.toHaveAttribute('data-swiping');
-      } finally {
-        document.elementFromPoint = originalElementFromPoint;
-        focusSpy.mockRestore();
-      }
-    },
-  );
-
-  it.skipIf(isJSDOM)('does not focus a keyboard input on touchend without touchstart', async () => {
-    await render(
-      <Drawer.Root open modal={false}>
-        <Drawer.VirtualKeyboardProvider>
-          <Drawer.Portal>
-            <Drawer.Viewport>
-              <Drawer.Popup>
-                <input data-testid="input" type="text" />
-              </Drawer.Popup>
-            </Drawer.Viewport>
-          </Drawer.Portal>
-        </Drawer.VirtualKeyboardProvider>
-      </Drawer.Root>,
-    );
-
-    const input = screen.getByTestId('input');
-    const focusSpy = vi.spyOn(input, 'focus');
-    const originalElementFromPoint = document.elementFromPoint;
-    document.elementFromPoint = () => input;
-
-    try {
-      const touchEnd = createNativeTouchEnd(input, {
-        clientX: 0,
-        clientY: 0,
-      });
-
-      await act(async () => {
-        input.dispatchEvent(touchEnd);
-        await flushMicrotasks();
-      });
-
-      expect(touchEnd.defaultPrevented).toBe(false);
-      expect(focusSpy).not.toHaveBeenCalled();
-    } finally {
-      document.elementFromPoint = originalElementFromPoint;
-      focusSpy.mockRestore();
-    }
-  });
-
-  it.skipIf(isJSDOM)('does not focus a keyboard input after touchcancel', async () => {
-    await render(
-      <Drawer.Root open modal={false}>
-        <Drawer.VirtualKeyboardProvider>
-          <Drawer.Portal>
-            <Drawer.Viewport>
-              <Drawer.Popup>
-                <input data-testid="input" type="text" />
-              </Drawer.Popup>
-            </Drawer.Viewport>
-          </Drawer.Portal>
-        </Drawer.VirtualKeyboardProvider>
-      </Drawer.Root>,
-    );
-
-    const input = screen.getByTestId('input');
-    const focusSpy = vi.spyOn(input, 'focus');
-    const originalElementFromPoint = document.elementFromPoint;
-    document.elementFromPoint = () => input;
-
-    try {
-      fireEvent.touchStart(input, {
-        touches: [
-          createTouch(input, {
-            clientX: 0,
-            clientY: 0,
-          }),
-        ],
-      });
-
-      const touchCancel = createNativeTouchCancel(input, {
-        clientX: 0,
-        clientY: 0,
-      });
-
-      await act(async () => {
-        input.dispatchEvent(touchCancel);
-        await flushMicrotasks();
-      });
-
-      const touchEnd = createNativeTouchEnd(input, {
-        clientX: 0,
-        clientY: 0,
-      });
-
-      await act(async () => {
-        input.dispatchEvent(touchEnd);
-        await flushMicrotasks();
-      });
-
-      expect(touchEnd.defaultPrevented).toBe(false);
-      expect(focusSpy).not.toHaveBeenCalled();
-    } finally {
-      document.elementFromPoint = originalElementFromPoint;
-      focusSpy.mockRestore();
-    }
-  });
-
-  it.skipIf(isJSDOM)('does not focus a keyboard input while a nested drawer is open', async () => {
-    await render(
-      <Drawer.Root open modal={false}>
-        <Drawer.VirtualKeyboardProvider>
-          <Drawer.Portal>
-            <Drawer.Viewport>
-              <Drawer.Popup data-testid="parent-popup">
-                <input data-testid="input" type="text" />
-                <Drawer.Root open modal={false}>
-                  <Drawer.Portal>
-                    <Drawer.Viewport>
-                      <Drawer.Popup>
-                        <button type="button">Nested action</button>
-                      </Drawer.Popup>
-                    </Drawer.Viewport>
-                  </Drawer.Portal>
-                </Drawer.Root>
-              </Drawer.Popup>
-            </Drawer.Viewport>
-          </Drawer.Portal>
-        </Drawer.VirtualKeyboardProvider>
-      </Drawer.Root>,
-    );
-
-    const input = screen.getByTestId('input');
-    const parentPopup = screen.getByTestId('parent-popup');
-    const focusSpy = vi.spyOn(input, 'focus');
-    const originalElementFromPoint = document.elementFromPoint;
-    document.elementFromPoint = () => input;
-
-    try {
-      await waitFor(() => {
-        expect(parentPopup).toHaveAttribute('data-nested-drawer-open', '');
-      });
-
-      fireEvent.touchStart(input, {
-        touches: [
-          createTouch(input, {
-            clientX: 0,
-            clientY: 0,
-          }),
-        ],
-      });
-
-      const touchEnd = createNativeTouchEnd(input, {
-        clientX: 0,
-        clientY: 0,
-      });
-
-      await act(async () => {
-        input.dispatchEvent(touchEnd);
-        await flushMicrotasks();
-      });
-
-      expect(touchEnd.defaultPrevented).toBe(false);
-      expect(focusSpy).not.toHaveBeenCalled();
-    } finally {
-      document.elementFromPoint = originalElementFromPoint;
-      focusSpy.mockRestore();
-    }
-  });
-
-  it.skipIf(isJSDOM)('preserves native taps on an already-focused keyboard input', async () => {
-    await render(
-      <Drawer.Root open modal={false}>
-        <Drawer.VirtualKeyboardProvider>
-          <Drawer.Portal>
-            <Drawer.Viewport>
-              <Drawer.Popup>
-                <input data-testid="input" type="text" />
-              </Drawer.Popup>
-            </Drawer.Viewport>
-          </Drawer.Portal>
-        </Drawer.VirtualKeyboardProvider>
-      </Drawer.Root>,
-    );
-
-    const input = screen.getByTestId('input');
-
-    await act(async () => {
-      input.focus();
-    });
-
-    const focusSpy = vi.spyOn(input, 'focus');
-    const originalElementFromPoint = document.elementFromPoint;
-    document.elementFromPoint = () => input;
-
-    try {
-      fireEvent.touchStart(input, {
-        touches: [
-          createTouch(input, {
-            clientX: 0,
-            clientY: 0,
-          }),
-        ],
-      });
-
-      const touchEnd = createNativeTouchEnd(input, {
-        clientX: 0,
-        clientY: 0,
-      });
-
-      await act(async () => {
-        input.dispatchEvent(touchEnd);
-        await flushMicrotasks();
-      });
-
-      expect(touchEnd.defaultPrevented).toBe(false);
-      expect(focusSpy).not.toHaveBeenCalled();
-    } finally {
-      document.elementFromPoint = originalElementFromPoint;
-      focusSpy.mockRestore();
     }
   });
